Updated gameplay and events
- Fixed seeking projectile (now rotates towards the closest enemy following an angular velocity)
- Fixed projectile raycasting (for collision detection)
- Fixed some effects: "Roll with the flow", "Just needed some grease", "Sacreblue! It's jammed again", "I went to the shooting range", "American Sniper" and "fear of damaging goods"

+++ Assets/Scripts/HUDManager.cs
@@ -100,7 +100,7 @@ public class HUDManager : MonoBehaviour
 		_damageLabel.text = $"Damage: <b>{ damage }</b>";
 	}
 
-	public void UpdateFireRateLabel(int fireRate)
+	public void UpdateFireRateLabel(float fireRate)
 	{
 		_fireRateLabel.text = $"Fire Rate: <b>{ fireRate } bursts/s</b>";
 	}
